Handles Next.js 15.3 Promise params
Addresses an issue introduced in Next.js 15.3 where `params` in page props can be a Promise.

This change unwraps the `params` promise to ensure the `id` is correctly accessed, preventing errors when fetching the apiary data. It also handles cases where the promise is undefined, preventing errors further down the line.